Split ChatArea send/mode logic into helpers (455 → 215 lines)
The send pipelines and conversation mutations are extracted out of
the shell so ChatArea is mostly just state + JSX wiring:

- chat-area/sendChatMessage.ts        SSE streaming chat-mode send,
                                      tempUser → token-by-token →
                                      committed turns, friendly error
                                      surfacing
- chat-area/sendAgentMessage.ts       agent-mode dispatch via
                                      /api/orchestrate, single-reply
                                      append with status-aware
                                      error messaging
- chat-area/useConversationMutations.ts  updateMode + rename + the
                                      first-message auto-title rule

Each helper takes its required state setters as args so the shell
stays the single owner of localMessages / streaming state. All four
files under 220 lines.
